requirements.txt
mypy>=0.610
pycryptodome==3.6.1

